javadoc fixes, var names fixes and some grammar fixes
