+++ Task_2_3_1/src/main/resources/ru/nsu/mbogdanov/fxml/menu.fxml
@@ -8,7 +8,7 @@
 
 <StackPane styleClass="background" prefHeight="500.0" prefWidth="500.0" stylesheets="@../css/menu.css"
            xmlns="http://javafx.com/javafx/16"
-           xmlns:fx="http://javafx.com/fxml/1" fx:controller="ru.nsu.mbogdanov.application.menu.MenuController">
+           xmlns:fx="http://javafx.com/fxml/1" fx:controller="ru.nsu.mbogdanov.controllers.windows.MenuController">
     <Rectangle styleClass="menu" height="400.0" width="400.0"/>
     <VBox maxHeight="400.0" maxWidth="400.0">
         <HBox alignment="CENTER" prefHeight="100.0" prefWidth="200.0">
