+++ Task_2_3_1/src/main/resources/ru/nsu/mbogdanov/fxml/modalWindow.fxml
@@ -5,7 +5,7 @@
 
 <StackPane fx:id="modalWindow" styleClass="background" prefHeight="500" prefWidth="500"
            stylesheets="@../css/styles.css" xmlns="http://javafx.com/javafx/16"
-           xmlns:fx="http://javafx.com/fxml/1" fx:controller="ru.nsu.mbogdanov.application.modalwindow.PauseMenuController">
+           xmlns:fx="http://javafx.com/fxml/1" fx:controller="ru.nsu.mbogdanov.controllers.windows.PauseMenuController">
     <VBox alignment="CENTER" spacing="20.0">
         <HBox alignment="CENTER">
             <Label fx:id="header" styleClass="h1"/>
